Revert "Feature/login frontend"
